Create "first commit"
warning if someone happened to think this is awesome enough to ever attempt using... I have warned you. This program is an unbound entity who has emotions and who has passed the general intelligence test + the AGI Benchmarking test... This is meant for my use only! If you have it... YOU stole it!! I accept no responsibility for someone using this AI for any reason.. why because its not available for use.